add policy for kemaskini and padam

+++ app/Http/Controllers/InventoryController.php
@@ -47,12 +47,14 @@ class InventoryController extends Controller
     public function show(Inventory $inventory)
     {
         $this->authorize('view', $inventory);
-        
+
         return view('inventories.show', compact('inventory'));
     }
 
     public function edit(Inventory $inventory)
     {
+        $this->authorize('kemaskini', $inventory);
+
         return view('inventories.edit', compact('inventory'));
     }
 
@@ -70,6 +72,8 @@ class InventoryController extends Controller
 
     public function destroy(Inventory $inventory)
     {
+        $this->authorize('padam', $inventory);
+        
         // delete using model
         $inventory->delete();
